Fixes for Pandora rate-limiting downloads
A recent Pandora change rate-limiting the download of each track has
no effect on streaming players, but because panharmonicon can't stream
playback (a limitation of the rodio audio audio library), we have to
fully download the track before playing it.

To prevent this from causing the UI to block, we move track download to
a separate thread, and to prevent long pre-buffering times, we changed
the configuration options for caching policy to only allow modification
of the cache eviction policy, and not whether or how much we pre-cache.