Multiple fixes for UI scaling and responsiveness
It turns out that at least one call to `siv.refresh()` needs to be made
before the first call to `siv.step()`, at least if you have a
WindowResize event callback that cares about the window size.  Otherwise
a size of 0,0 will be reported and wrong decisions are made, which won't
be corrected until the first resize event after starting.  This fixes
the UI starting at the top of the window, then dropping to the bottom
after resizing.

When idling, poll input occasionally while running out the heartbeat
timer. This makes the UI more responsive to user input when nothing else
is going on.  Otherwise, the input handling might be delayed until the
heartbeat timer runs out.