Updates to CI/CD for binary stripping, debug symbol handling, build caching
macOS builds broke because of something around debug symbol stripping.
In response, we're switching to using rust's own stripping via
Cargo.toml, and since we don't get a .debug file anymore as a result,
and because it wasn't seeing much use, we're also eliminating packaging
and releasing debug symbols for releases.

At the same time, taking advantage of actions for improved build
caching, and more explicitly managing the build toolchain.