Retry failed api requests
Sessions can timeout, and going through user_login() before each API
call doesn't ensure that a just-expired session will get logged in.  It
will only take care of the case where a previous request returned
failure due to an expired session, so we login before trying the next
request.

So instead, we'll try each request exactly as-is, and if it fails (for
any reason) we'll try a login, then retry the request.

This should have the benefit of eliminating some overhead from every API
call that would succeed, all the while ensuring that more kinds of
failures get turned into successes.