Switch to only awaiting subsytems when they have work to do
With join!(), we were blocking the whole program whenever one subsystem
had a long step to run. This change switches to select!{}, which will
poll workers until one is ready, and then run that one.

The problem is, the other workers are cancelled, so they don't get to do
any work, so if one subsystem comes back quickly, the others are
prevented from doing work.

To mitigate this, we add a ready() method to each subsystem, so we're
only running subsystems that say that they have work to do (preventing a
subsytem with no work to do getting in the way of others that do).  For
the model subsystem, though, it's difficult to comprehensively determine
whether it has work to do - it's not just completing events off an event
queue - it's what's filling event queues.  So when all the event queues
across all subsystems are empty, and thus all the ready() methods return
empty, we'll sleep a bit (to prevent spinning in a tight loop when
nothing's going on), and then run the model for one tick.

This has the effect of providing much better responsiveness, without
negatively affecting CPU utilization.