Got task starvation, switched task scheduling model back
The issue with the old task scheduling model was basically that it would
tick each subsytem once before ticking each one again.  This was a
problem because the a tick of the caching subsystem could encompass
downloading an entire track, which wasn't great for responsiveness.

The new scheduling model would randomly schedule whichever subsystem is
ready, but this seemed to, under unknown but circumstantially very
repeatable conditions, result in the ui subsystem starving out the other
subsystems.

So as part of going back to the old scheduling model, the caching
subsystem was updating to do the actual fetching in a separate thread.
This makes each subsystem tick pretty short, as it only needs to check
each fetch thread to see if it has exited.