Massive refactor, probably very broken still
This refactor is built around changes to the messaging between
subsystems.  The model receives requests from other parts of the
system, and sends out state updates about what's changed in response to
the requests.  Consequently, the requests sent to the model are a
multiple-producer-single-consumer type of channel, while the state
updates from the model to all the subsystems are a broadcast model.

In order to make the abstractions cleaner and clearer, all audio control
logic (decoding a track and playing it, volume, muting, pausing) were
moved out into a separate subsystem.  This seems to create a much
cleaner separation of concerns, making it much easier to reason about
what's going on in the code being looked at.