Make the set of active fetch tasks wide, and extend timeout
There's a surprising number of fetches that end in failure due to
timeout on multiple retries, but the 10s timeout is too short for a
number of tracks that do succeed.  Even at 30s timeout, we still get
some tracks finishing in success after 29s.  In order to ensure a
smoother pipeline of playable tracks, we lengthen the timeout to 30s,
while allowing more simultaneous active fetch requests.

Even with a very low rate of pre-cached tracks, this should result in a
_relatively_ smooth experience, with some hitches early in the session
when it hasn't yet had much time to pre-cache much.